% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DELE_function.R
\name{DELE_function}
\alias{DELE_function}
\title{DELE_function}
\usage{
DELE_function(data, DELEcode, var, mapTitle = "", mapColor = NULL,
  titleColor = "black", showVar1, showVar2, showVar3, addLeg = T,
  prov = providers$Esri.WorldImagery, GovBoundaryColor = "black")
}
\arguments{
\item{data}{a data set that contains at least your states and the variable to represent}

\item{var}{a numeric that indicate the colum number that contains  your variable}

\item{mapTitle}{the map title}

\item{mapColor}{the map color}

\item{titleColor}{the color for the title}

\item{showVar1}{: the variables you want to show in the highlight box on the map}

\item{showVar2}{: the variables you want to show in the highlight box on the map}

\item{showVar3}{: the variables you want to show in the highlight box on the map}

\item{addLeg}{: a boolean to set to true in you want to have legend by the side of the map}

\item{prov}{: it have to be in format : providers$... and the provider you want to have in the background you can see all the providers by typing View(providers) or going to https://leaflet-extras.github.io/leaflet-providers/preview/ where you can fin all the providers and how they look like}

\item{GovBoundaryColor}{: the color for the states boundaries}

\item{HASC_2}{a numeric that indicate the colum number that contains the Delegations code}
}
\value{
a map with your data
}
\description{
mapping data by delegation
}
\examples{

testD <- cbind(tnMAPDELE@data[["NAME_2"]],tnMAPDELE@data[["HASC_2"]],tnMAPDELE@data[["NAME_1"]],rnorm(268,50,20))
testD=as.data.frame(testD)
testD[,4]=as.numeric(testD[,4])
DELE_function(data = testD,var=4,mapTitle= "new MAp ",titleColor = "red",mapColor = NULL, HASC_2 =   2,showVar1 = 1,showVar2 = 4,showVar3 = 3,addLeg = T,prov= providers$Esri.OceanBasemap,GovBoundaryColor="black")

}
