% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DELEDotDensityMap.R
\name{DELEDotDensityMap}
\alias{DELEDotDensityMap}
\title{DELEDotDensityMap}
\usage{
DELEDotDensityMap(data, DELEcode, var, DotWeight = 0.5,
  polyEdgColor = "blue", OneDFor = 50,
  dotColor = "lightgoldenrodyellow",
  prov = providers$NASAGIBS.ViirsEarthAtNight2012,
  polyColor = "midnightblue")
}
\arguments{
\item{data}{a data set that contains at least your delegations  and the variable to represent}

\item{DELEcode}{a numeric that indicate the colum number that contains the delegations code}

\item{var}{a numeric that indicate the colum number that contains  your variable}

\item{DotWeight}{the dot weight}

\item{polyEdgColor}{the boundries color}

\item{OneDFor}{the unit (exp: 60 to present 60 votes per dot )}

\item{dotColor}{the dots color}

\item{prov}{: it have to be in format : providers$... and the provider you want to have in the background you can see all the providers by typing View(providers) or going to https://leaflet-extras.github.io/leaflet-providers/preview/ where you can fin all the providers and how they look like}

\item{polyColor}{the polygons filling color}
}
\value{
a map with your data
}
\description{
DotDensity mapping data by state
}
\examples{
testD <- cbind(tnMAPDELE@data[["NAME_2"]],tnMAPDELE@data[["HASC_2"]],tnMAPDELE@data[["NAME_1"]],rnorm(268,50,20))
testD=as.data.frame(testD)
testD[,4]=as.numeric(testD[,4])
DELEDotDensityMap(testD,DELEcode = 2, polyEdgColor= "red" ,var=4,OneDFor = 60)


}
