% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gov_function.R
\name{Gov_function}
\alias{Gov_function}
\title{Gov_function
mapping data by state}
\usage{
Gov_function(data, var, mapTitle = "", mapColor = NULL,
  titleColor = "black", GOVcode, showVar1, showVar2, addLeg = T,
  prov = providers$Esri.WorldImagery)
}
\arguments{
\item{data}{a data set that contains at least your states and the variable to represent}

\item{var}{a numeric that indicate the colum number that contains  your variable}

\item{mapTitle}{the map title}

\item{mapColor}{the map color}

\item{titleColor}{the color for the title}

\item{GOVcode}{a numeric that indicate the colum number that contains the states code}

\item{showVar1}{: the variables you want to show in the highlight box on the map}

\item{showVar2}{: the variables you want to show in the highlight box on the map}

\item{addLeg}{: a boolean to set to true in you want to have legend by the side of the map}

\item{prov}{: it have to be in format : providers$... and the provider you want to have in the background you can see all the providers by typing View(providers) or going to https://leaflet-extras.github.io/leaflet-providers/preview/ where you can fin all the providers and how they look like}
}
\value{
a map with your data
}
\description{
Gov_function
mapping data by state
}
\examples{

testG <- cbind(tnMAP$NAME_1,rnorm(n = 24,mean = 20,sd = 13),tnMAP$HASC_1)
testG=as.data.frame(testG)
testG[,2]=as.numeric(testG[,2])
Gov_function(data = testG,var=2,mapTitle= "new MAp ",titleColor = "red",mapColor =c("white","red"),GOVcode =  3,addLeg = T,prov = providers$OpenStreetMap,showVar1 = 1,showVar2 = 2)

}
