% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GovDotDensityMap.R
\name{GovDotDensityMap}
\alias{GovDotDensityMap}
\title{DELEDotDensityMap}
\usage{
GovDotDensityMap(data, Govcode, var, DotWeight = 0.5,
  polyEdgColor = "blue", OneDFor = 50,
  dotColor = "lightgoldenrodyellow",
  prov = providers$NASAGIBS.ViirsEarthAtNight2012,
  polyColor = "midnightblue")
}
\arguments{
\item{data}{a data set that contains at least your states  and the variable to represent}

\item{Govcode}{a numeric that indicate the colum number that contains the states code}

\item{var}{a numeric that indicate the colum number that contains  your variable}

\item{DotWeight}{the dot weight}

\item{polyEdgColor}{the boundries color}

\item{OneDFor}{the unit (exp: 60 to present 60 votes per dot )}

\item{dotColor}{the dots color}

\item{prov}{: it have to be in format : providers$... and the provider you want to have in the background you can see all the providers by typing View(providers) or going to https://leaflet-extras.github.io/leaflet-providers/preview/ where you can fin all the providers and how they look like}

\item{polyColor}{the polygons filling color}
}
\value{
a map with your data
}
\description{
DELEDotDensityMap
}
\examples{
testG <- cbind(tnMAP$NAME_1,rnorm(n = 24,mean = 20,sd = 13),tnMAP$HASC_1)
testG=as.data.frame(testG)
testG[,2]=as.numeric(testG[,2])
GovDotDensityMap(testG,Govcode = 3,var=2,prov = providers$OpenStreetMap)

}
